fix(OY2-25752): os export hide columns
