+++ src/packages/shared-types/opensearch/_.ts
@@ -80,9 +80,4 @@ export type AggResult = Record<
   }
 >;
 
-export type ExportHeaderOptions<TData> = {
-  transform: (data: TData) => string;
-  name: string;
-};
-
 export type Index = "main" | "insights" | "changelog" | "types" | "subtypes";
