+++ src/packages/shared-types/opensearch/main/index.ts
@@ -4,7 +4,6 @@ import {
   Filterable as FIL,
   QueryState,
   AggQuery,
-  ExportHeaderOptions,
 } from "./../_";
 import { z } from "zod";
 import { ItemResult as Changelog } from "./../changelog";
@@ -38,6 +37,5 @@ export type Field = keyof Document | `${keyof Document}.keyword`;
 export type Filterable = FIL<Field>;
 export type State = QueryState<Field>;
 export type Aggs = AggQuery<Field>;
-export type ExportHeader = ExportHeaderOptions<Document>;
 
 export * from "./transforms";
