+++ src/packages/shared-types/opensearch/subtypes/index.ts
@@ -4,7 +4,6 @@ import {
   Filterable as FIL,
   QueryState,
   AggQuery,
-  ExportHeaderOptions,
 } from "./../_";
 import { z } from "zod";
 import { Type } from "./transforms";
@@ -20,6 +19,5 @@ export type Field = keyof Document | `${keyof Document}.keyword`;
 export type Filterable = FIL<Field>;
 export type State = QueryState<Field>;
 export type Aggs = AggQuery<Field>;
-export type ExportHeader = ExportHeaderOptions<Document>;
 
 export * from "./transforms";
