+++ src/services/ui/src/components/Opensearch/main/Table/types.ts
@@ -4,9 +4,10 @@ import type { ReactNode } from "react";
 export type OsTableColumn = {
   field?: opensearch.main.Field;
   label: string;
-  visible?: boolean;
   locked?: boolean;
+  hidden?: boolean;
   isSystem?: boolean;
   props?: any;
+  transform?: (data: opensearch.main.Document) => string;
   cell: (data: opensearch.main.Document) => ReactNode;
 };
