--- src/services/ui/src/components/Opensearch/main/index.ts
@@ -1,6 +1,0 @@
-export * from "./useOpensearch";
-export * from "./types";
-export * from "./Table";
-export * from "./Filtering";
-export * from "./Provider";
-export * from "./Settings";
