+++ src/services/ui/src/components/index.tsx
@@ -9,7 +9,6 @@ export * from "./Context";
 export * from "./DetailsSection";
 export * from "./Dialog";
 export * from "./ErrorAlert";
-export * from "./ExportButton";
 export * from "./Footer";
 export * from "./Form";
 export * from "./GovernmentBuildingIcon";
