+++ src/services/ui/src/features/dashboard/Lists/spas/index.tsx
@@ -1,41 +1,11 @@
-import {
-  OsTable,
-  OsFiltering,
-  useOsContext,
-  useOsUrl,
-  Pagination,
-  ErrorAlert,
-} from "@/components";
+import { useOsContext, ErrorAlert, OsMainView } from "@/components";
 import { useSpaTableColumns } from "./consts";
 
 export const SpasList = () => {
   const context = useOsContext();
-  const url = useOsUrl();
   const columns = useSpaTableColumns();
 
   if (context.error) return <ErrorAlert error={context.error} />;
 
-  return (
-    <section className="flex flex-col h-[100vh]">
-      <OsFiltering />
-      <OsTable columns={columns} />
-      <Pagination
-        pageNumber={url.state.pagination.number}
-        count={context.data?.total?.value || 0}
-        pageSize={url.state.pagination.size}
-        onPageChange={(number) =>
-          url.onSet((s) => ({
-            ...s,
-            pagination: { ...s.pagination, number },
-          }))
-        }
-        onSizeChange={(size) =>
-          url.onSet((s) => ({
-            ...s,
-            pagination: { number: 0, size },
-          }))
-        }
-      />
-    </section>
-  );
+  return <OsMainView columns={columns} />;
 };
