+++ src/services/ui/src/features/dashboard/Lists/waivers/consts.tsx
@@ -8,6 +8,7 @@ import {
   renderCellDate,
   renderCellIdLink,
 } from "../renderCells";
+import { formatSeatoolDate } from "shared-utils";
 
 export const useWaiverTableColumns = (): OsTableColumn[] => {
   const { data: props } = useGetUser();
@@ -20,17 +21,19 @@ export const useWaiverTableColumns = (): OsTableColumn[] => {
       field: "id.keyword",
       label: "Waiver Number",
       locked: true,
+      transform: (data) => data.id,
       cell: renderCellIdLink((id) => `/details?id=${encodeURIComponent(id)}`),
     },
     {
       field: "state.keyword",
       label: "State",
-      visible: true,
+      transform: (data) => data.state ?? BLANK_VALUE,
       cell: (data) => data.state,
     },
     {
       field: "authority.keyword",
       label: "Type",
+      transform: (data) => data.authority ?? BLANK_VALUE,
       cell: (data) =>
         data?.authority
           ? removeUnderscoresAndCapitalize(data.authority)
@@ -39,6 +42,12 @@ export const useWaiverTableColumns = (): OsTableColumn[] => {
     {
       field: "actionType.keyword",
       label: "Action Type",
+      transform: (data) => {
+        if (data.actionType === undefined) return BLANK_VALUE;
+        return (
+          LABELS[data.actionType as keyof typeof LABELS] || data.actionType
+        );
+      },
       cell: (data) =>
         data.actionType
           ? LABELS[data.actionType as keyof typeof LABELS] || data.actionType
@@ -47,6 +56,27 @@ export const useWaiverTableColumns = (): OsTableColumn[] => {
     {
       field: props?.isCms ? "cmsStatus.keyword" : "stateStatus.keyword",
       label: "Status",
+      transform: (data) => {
+        const status = (() => {
+          if (!props?.isCms) return data.stateStatus;
+          if (props?.user?.["custom:cms-roles"].includes(UserRoles.HELPDESK)) {
+            return data.stateStatus;
+          }
+          return data.cmsStatus;
+        })();
+
+        const subStatusRAI = data.raiWithdrawEnabled
+          ? " (Withdraw Formal RAI Response - Enabled)"
+          : "";
+
+        const subStatusInitialIntake = (() => {
+          if (!props?.isCms) return "";
+          if (!data.initialIntakeNeeded) return "";
+          return " (Initial Intake Needed)";
+        })();
+
+        return `${status}${subStatusRAI}${subStatusInitialIntake}`;
+      },
       cell: (data) => {
         const status = (() => {
           if (!props?.isCms) return data.stateStatus;
@@ -73,12 +103,16 @@ export const useWaiverTableColumns = (): OsTableColumn[] => {
     {
       field: "submissionDate",
       label: "Initial Submission",
+      transform: (data) =>
+        data?.submissionDate
+          ? formatSeatoolDate(data.submissionDate)
+          : BLANK_VALUE,
       cell: renderCellDate("submissionDate"),
     },
     {
       field: "origin",
       label: "Submission Source",
-      visible: false,
+      hidden: true,
       cell: (data) => {
         return data.origin;
       },
@@ -86,23 +120,30 @@ export const useWaiverTableColumns = (): OsTableColumn[] => {
     {
       field: "raiRequestedDate",
       label: "Formal RAI Requested",
-      visible: false,
+      hidden: true,
       cell: renderCellDate("raiRequestedDate"),
     },
     {
       field: "raiReceivedDate",
       label: "Formal RAI Response",
+      transform: (data) => {
+        return data.raiReceivedDate && !data.raiWithdrawnDate
+          ? formatSeatoolDate(data.raiReceivedDate)
+          : BLANK_VALUE;
+      },
       cell: renderCellDate("raiReceivedDate"),
     },
     {
       field: "leadAnalystName.keyword",
       label: "CPOC Name",
-      visible: false,
+      hidden: true,
+      transform: (data) => data.leadAnalystName ?? BLANK_VALUE,
       cell: (data) => data.leadAnalystName,
     },
     {
       field: "submitterName.keyword",
       label: "Submitted By",
+      transform: (data) => data.submitterName ?? BLANK_VALUE,
       cell: (data) => data.submitterName,
     },
     // hide actions column for: readonly,help desk
